blade: only install non-essential files if system partition is large enough
Customised partition layouts are common on blade. Although the default
layout used by ZTE has a large enough system partition to install the
full CM9 rom many users prefer to use a smaller system partition to
allow for a larger data partition (the entire nand is only 512MB).

This patch allows for non-essential content to be installed only if
the system partition is large enough. Content added to extras.txt will
only be installed if the system partition is 210MB or larger.